Ref: size型(long long unsigned int)と比較のある場合カウンタ変数をsize_tに変更

+++ introduction/model_answer/12_takahashikun_and_string_operation.cpp
@@ -8,7 +8,7 @@ int main() {
 	string ss = s + s;
 	auto rot_s_left = ss.end() - s.length();
 	auto rot_s_right = ss.end();
-	for (int i = 0; i < s.size(); i++) {
+	for (size_t i = 0; i < s.size(); i++) {
 		if (equal(t.begin(), t.end(), rot_s_left, rot_s_right)) {
 			ans = i;
 			break;
